move chain table names to other dir
