--- util/constants.js
@@ -1,7 +1,0 @@
-const chainTableNames = {
-  rsk_testnet: 'chain_rsk_testnet',
-  rsk_mainnet: 'chain_rsk_mainnet',
-};
-module.exports = {
-  chainTableNames,
-};
